# docs/rest/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreatePVZIn:
    properties:
      city:
        type: string
    required:
    - city
    type: object
  dto.CreatePVZOut:
    properties:
      city:
        type: string
      id:
        type: string
      registrationDate:
        type: string
    type: object
...
  dto.PvzInfoIn:
    properties:
      endDate:
        type: string
      limit:
        type: integer
      page:
        type: integer
      startDate:
        type: string
    required:
    - endDate
    - limit
    - page
    - startDate
    type: object
  dto.PvzInfoOut:
    properties:
      pvz:
        $ref: '#/definitions/dto.PvzInfoPvzOut'
      receptions:
        items:
          $ref: '#/definitions/dto.PvzInfoReceptionWithProductsOut'
        type: array
    type: object
  dto.PvzInfoProductOut:
    properties:
      dateTime:
        type: string
      id:
        type: string
      receptionId:
        type: string
      type:
        type: string
    type: object
  dto.PvzInfoPvzOut:
    properties:
      city:
        type: string
      id:
        type: string
      registrationDate:
        type: string
    type: object
  dto.PvzInfoReceptionOut:
    properties:
      dateTime:
        type: string
      id:
        type: string
      pvzId:
        type: string
      status:
        type: string
    type: object
  dto.PvzInfoReceptionWithProductsOut:
    properties:
      products:
        items:
          $ref: '#/definitions/dto.PvzInfoProductOut'
        type: array
      reception:
        $ref: '#/definitions/dto.PvzInfoReceptionOut'
    type: object
  dto.RegisterUserIn:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        type: string
    required:
    - email
    - password
    - role
    type: object
  dto.RegisterUserOut:
    properties:
      email:
        type: string
      id:
        type: string
      role:
        type: string
    type: object
  dto.StartReceptionIn:
    properties:
      pvzId:
        type: string
    required:
    - pvzId
    type: object
  dto.StartReceptionOut:
    properties:
      datetime:
        type: string
      id:
        type: string
      pvzId:
        type: string
      status:
        type: string
    type: object
  handler.errorResponse:
    properties:
      details:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a service for working with PVZ.
  title: PVZ service
  version: "1.0"
paths:
...
  /pvz:
    get:
      consumes:
      - application/json
      description: Get PVZ list with receptions and products in given time period
      operationId: GetPVZInfo
      parameters:
      - description: Time period and pagination parameters
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.PvzInfoIn'
      produces:
      - application/json
      responses:
        "200":
          description: List of PVZs with receptions and products
          schema:
            items:
              $ref: '#/definitions/dto.PvzInfoOut'
            type: array
        "400":
          description: Invalid parameters or no data found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get PVZ information
      tags:
      - PVZ
    post:
      consumes:
      - application/json
      description: Create new PVZ. Requires JWT-Token with Employee role.
      operationId: CreatePVZ
      parameters:
      - description: PVZ creation data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePVZIn'
      produces:
      - application/json
      responses:
        "200":
          description: PVZ successfully created
          schema:
            $ref: '#/definitions/dto.CreatePVZOut'
        "400":
          description: Invalid request or validation failed
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create PVZ
      tags:
      - PVZ
...
  /receptions:
    post:
      consumes:
      - application/json
      description: Create new opened reception at PVZ
      operationId: StartReception
      parameters:
      - description: PVZ ID
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.StartReceptionIn'
      produces:
      - application/json
      responses:
        "200":
          description: Reception successfully started
          schema:
            $ref: '#/definitions/dto.StartReceptionOut'
        "400":
          description: PVZ not found or opened reception exists
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Start new reception
      tags:
      - PVZ
  /register:
    post:
      consumes:
      - application/json
      description: Create new user account
      operationId: RegisterUser
      parameters:
      - description: User registration data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterUserIn'
      produces:
      - application/json
      responses:
        "200":
          description: User successfully registered
          schema:
            $ref: '#/definitions/dto.RegisterUserOut'
        "400":
          description: Validation failed or user already exists
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Register new user
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"